fix: cross-position attention in JAX and PyTorch inference
jax.nn.dot_product_attention(q, k, v, is_causal=True) with layout
(B, H, T, D) computes attention across heads, not positions. Replaced
with a manual einsum and fp32 softmax (needed for numerical stability
with NativeBit quantized weights).

Verified empirically via a context-sensitivity probe:
  KL(predict(full_context), predict(single_token))
  before fix: 0.0001  (model ignores context)
  after fix:  0.18    (context matters)

PyTorch inference updated to match the new attention semantics and
switched to bf16 (weights were trained in bf16 on TPU; fp16 overflowed
during inference on some layers).

Tests directory is now tracked (removed blanket test_*.py ignore,
kept root-level and experiments/ test scripts private).

+++ inference/eval_ppl.py
@@ -163,14 +163,14 @@ def main():
             block_idx = torch.arange(num_blocks).unsqueeze(1)
             total = module.out_features * module.in_features
             w = codebook[block_idx, indices].reshape(-1)[:total]
-            w = w.reshape(module.out_features, module.in_features).half()
+            w = w.reshape(module.out_features, module.in_features).to(torch.bfloat16)
             linear = torch.nn.Linear(module.in_features, module.out_features, bias=False)
             linear.weight.data = w
             parts = name.rsplit(".", 1)
             if len(parts) == 2:
                 parent = dict(nb_model.named_modules())[parts[0]]
                 setattr(parent, parts[1], linear)
-    nb_model.half()
+    nb_model.to(torch.bfloat16)
     nb_model.to(DEVICE)
     vram_mb = torch.cuda.memory_allocated() / 1e6
     print(f"  Done. VRAM: {vram_mb:.0f} MB")
